Added piping support to writeable streams
Currently the library was buffering the response before we’re able to send the response anywhere.

Since I needed a highly-concurrent use of this, I patched the code to scratch my own itch :)

Happy to modify it, if you feel like there’s a better way of adding this feature.